Fix bugs in AMSDEBUG instrumentation.

+++ ams/library/dgrts.c
@@ -69,7 +69,7 @@ static int	dgrMamsInit(MamsInterface *tsif)
 	}
 
 #if AMSDEBUG
-printf("parsed endpoint spec to port %d address %d.\n", portNbr, ipAddress);
+printf("parsed endpoint spec to port %hu address %u.\n", portNbr, ipAddress);
 #endif
 	if (dgr_open(sm_TaskIdSelf(), DGRTS_CLIENT_SVC_ID, portNbr, ipAddress,
 			memmgr_name(getIonMemoryMgr()), &dgrSap, &rc) < 0)
@@ -282,7 +282,7 @@ static int	dgrParseMamsEndpoint(MamsEndpoint *ep)
 	CHKERR(ep->tsep);
 	memcpy((char *) (ep->tsep), (char *) &tsep, sizeof(DgrTsep));
 #if AMSDEBUG
-printf("parsed '%s' to port %d address %d.\n", ep->ept, tsep.portNbr,
+printf("parsed '%s' to port %hu address %u.\n", ep->ept, tsep.portNbr,
 tsep.ipAddress);
 #endif
 	return 0;
@@ -341,8 +341,8 @@ static int	dgrSendMams(MamsEndpoint *ep, MamsInterface *tsif, char *msg,
 	CHKERR(msgLen >= 0);
 	tsep = (DgrTsep *) (ep->tsep);
 #if AMSDEBUG
-printf("in dgrSendMams, tsep at %d has port %d, address %d.\n", (int) tsep,
-tsep->portNbr, tsep->ipAddress);
+printf("in dgrSendMams, tsep at %lu has port %hu, address %u.\n",
+(unsigned long) tsep, tsep->portNbr, tsep->ipAddress);
 #endif
 	if (tsep == NULL)	/*	Lost connectivity to endpoint.	*/
 	{
@@ -388,7 +388,7 @@ static int	dgrSendAms(AmsEndpoint *dp, AmsSAP *sap,
 	CHKERR(len <= DGRTS_MAX_MSG_LEN);
 	tsep = (DgrTsep *) (dp->tsep);
 #if AMSDEBUG
-printf("in dgrSendAms, tsep is %d.\n", (int) tsep);
+printf("in dgrSendAms, tsep is %lu.\n", (unsigned long) tsep);
 #endif
 	if (tsep == NULL)	/*	Lost connectivity to endpoint.	*/
 	{
